+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod capturer;
-pub mod duplicate_context;
+pub mod duplication_context;
 pub mod manager;
 pub mod utils;