+++ src/capturer/custom.rs
@@ -3,7 +3,7 @@ use windows::Win32::Graphics::{
   Dxgi::{DXGI_OUTDUPL_FRAME_INFO, DXGI_OUTDUPL_POINTER_SHAPE_INFO, DXGI_OUTPUT_DESC},
 };
 
-use crate::utils::Result;
+use crate::model::Result;
 use crate::{duplication_context::DuplicationContext, utils::OutputDescExt};
 
 use super::model::Capturer;