+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod capturer;
 pub mod duplication_context;
 pub mod manager;
+pub mod model;
 pub mod utils;