+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 - **_Breaking Change_**: error type changed from `&static str` to `String`.
 - **_Breaking Change_**: rename `DXGI_OUTDUPL_FRAME_INFO.is_new_frame` to `DXGI_OUTDUPL_FRAME_INFO.desktop_updated`, add `DXGI_OUTDUPL_FRAME_INFO.mouse_updated`.
+- **_Breaking Change_**: rename `DuplicateContext` to `DuplicationContext`.
 - Feat: add `CustomCapturer`.
 - Feat: add `SharedCapturer.open`.
-- Feat: add `DuplicateContext.get_pointer`.
+- Feat: add `DuplicationContext.get_pointer`.
 
 ## v0.2.0