+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v0.4.0
+
+- **_Breaking Change_**: move `Result` to `model` module.
+
 ## v0.3.0
 
 - **_Breaking Change_**: error type changed from `&static str` to `String`.